Move stars logic into dedicated controller package
Refactor GitHub stars management code from main.go into a new StarsCtl
controller in the internal/starsctl package. This change improves code
organization and separates concerns by:
- Creating a new StarsCtl type to manage stars operations
- Moving database initialization into the controller
- Adding proper error handling and context support
- Fetching the current user instead of hardcoding username
